Separate tss tests into separate packages to make them more reproducible
