+++ tss/signing/signing_test.go
@@ -1,0 +1,141 @@
+package signing_test
+
+import (
+	"context"
+	"fmt"
+	"math/big"
+	"time"
+
+	"github.com/ChainSafe/sygma-relayer/comm"
+	"github.com/ChainSafe/sygma-relayer/comm/elector"
+	"github.com/ChainSafe/sygma-relayer/keyshare"
+	"github.com/ChainSafe/sygma-relayer/tss"
+	"github.com/ChainSafe/sygma-relayer/tss/keygen"
+	"github.com/ChainSafe/sygma-relayer/tss/signing"
+	tsstest "github.com/ChainSafe/sygma-relayer/tss/test"
+	"github.com/libp2p/go-libp2p-core/peer"
+)
+
+type SigningTestSuite struct {
+	tsstest.CoordinatorTestSuite
+}
+
+func (s *SigningTestSuite) Test_ValidSigningProcess() {
+	communicationMap := make(map[peer.ID]*tsstest.TestCommunication)
+	coordinators := []*tss.Coordinator{}
+	processes := []tss.TssProcess{}
+
+	for i, host := range s.Hosts {
+		communication := tsstest.TestCommunication{
+			Host:          host,
+			Subscriptions: make(map[comm.SubscriptionID]chan *comm.WrappedMessage),
+		}
+		communicationMap[host.ID()] = &communication
+		fetcher := keyshare.NewKeyshareStore(fmt.Sprintf("../test/keyshares/%d.keyshare", i))
+
+		msgBytes := []byte("Message")
+		msg := big.NewInt(0)
+		msg.SetBytes(msgBytes)
+		signing, err := signing.NewSigning(msg, "signing1", host, &communication, fetcher)
+		if err != nil {
+			panic(err)
+		}
+		electorFactory := elector.NewCoordinatorElectorFactory(host, s.BullyConfig)
+		coordinators = append(coordinators, tss.NewCoordinator(host, &communication, electorFactory))
+		processes = append(processes, signing)
+	}
+	tsstest.SetupCommunication(communicationMap)
+
+	statusChn := make(chan error, s.PartyNumber)
+	resultChn := make(chan interface{})
+	ctx, cancel := context.WithCancel(context.Background())
+	for i, coordinator := range coordinators {
+		go coordinator.Execute(ctx, processes[i], resultChn, statusChn)
+	}
+
+	err := <-statusChn
+	s.Nil(err)
+	sig := <-resultChn
+	s.NotNil(sig)
+	time.Sleep(time.Millisecond * 50)
+	cancel()
+}
+
+func (s *SigningTestSuite) Test_SigningTimeout() {
+	communicationMap := make(map[peer.ID]*tsstest.TestCommunication)
+	coordinators := []*tss.Coordinator{}
+	processes := []tss.TssProcess{}
+
+	for i, host := range s.Hosts {
+		communication := tsstest.TestCommunication{
+			Host:          host,
+			Subscriptions: make(map[comm.SubscriptionID]chan *comm.WrappedMessage),
+		}
+		communicationMap[host.ID()] = &communication
+		fetcher := keyshare.NewKeyshareStore(fmt.Sprintf("../test/keyshares/%d.keyshare", i))
+
+		msgBytes := []byte("Message")
+		msg := big.NewInt(0)
+		msg.SetBytes(msgBytes)
+		signing, err := signing.NewSigning(msg, "signing2", host, &communication, fetcher)
+		if err != nil {
+			panic(err)
+		}
+		electorFactory := elector.NewCoordinatorElectorFactory(host, s.BullyConfig)
+		coordinator := tss.NewCoordinator(host, &communication, electorFactory)
+		coordinator.TssTimeout = time.Nanosecond
+		coordinators = append(coordinators, coordinator)
+		processes = append(processes, signing)
+	}
+	tsstest.SetupCommunication(communicationMap)
+
+	statusChn := make(chan error, s.PartyNumber)
+	resultChn := make(chan interface{})
+	ctx, cancel := context.WithCancel(context.Background())
+	for i, coordinator := range coordinators {
+		go coordinator.Execute(ctx, processes[i], resultChn, statusChn)
+	}
+
+	err := <-statusChn
+	s.NotNil(err)
+	err = <-statusChn
+	s.NotNil(err)
+	err = <-statusChn
+	s.NotNil(err)
+	time.Sleep(time.Millisecond * 50)
+	cancel()
+}
+
+func (s *SigningTestSuite) Test_PendingProcessExists() {
+	communicationMap := make(map[peer.ID]*tsstest.TestCommunication)
+	coordinators := []*tss.Coordinator{}
+	processes := []tss.TssProcess{}
+	for _, host := range s.Hosts {
+		communication := tsstest.TestCommunication{
+			Host:          host,
+			Subscriptions: make(map[comm.SubscriptionID]chan *comm.WrappedMessage),
+		}
+		communicationMap[host.ID()] = &communication
+		keygen := keygen.NewKeygen("keygen3", s.Threshold, host, &communication, s.MockStorer)
+		electorFactory := elector.NewCoordinatorElectorFactory(host, s.BullyConfig)
+		coordinators = append(coordinators, tss.NewCoordinator(host, &communication, electorFactory))
+		processes = append(processes, keygen)
+	}
+	tsstest.SetupCommunication(communicationMap)
+
+	s.MockStorer.EXPECT().LockKeyshare().AnyTimes()
+	status := make(chan error, s.PartyNumber)
+	ctx, cancel := context.WithCancel(context.Background())
+	for i, coordinator := range coordinators {
+		go coordinator.Execute(ctx, processes[i], nil, nil)
+		time.Sleep(time.Millisecond * 50)
+		go coordinator.Execute(ctx, processes[i], nil, status)
+	}
+
+	for i := 0; i < s.PartyNumber; i++ {
+		err := <-status
+		s.Nil(err)
+	}
+	time.Sleep(time.Millisecond * 50)
+	cancel()
+}
